requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
argon2-cffi>=21.1.0
APScheduler>=3.8
psycopg2-binary>=2.9
python-dotenv>=0.19.0
